+++ resource-registry/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Revision history of `resource-registry`
 
+## Next version
+
+* Define new functions that implement tracing via `contra-tracer` package:
+  `bracketWithPrivateTracingRegistry`, `withTracingRegistry`,
+  `allocateLabelledTemp`, `modifyWithTempTracingRegistry`,
+  `runInnerWithTempTracingRegistry`, `runWithTempTracingRegistry`,
+  `unsafeNewRegistryWithTracer`, `allocateLabelled`, `allocateLabelledThread`,
+  `allocateEitherLabelled`.
+
 ## 0.2.0.0 — 2025-10-23
 
 * Define `transferRegistry` for moving all resources from one registry to a